compiler/libec: (#341, #351, #644, #771) Improved enum type matching and type handling of operator expressions ecere, ide, eda, samples: Fixed related warnings - Solved enum types popping up in auto-complete when they shouldn't, e.g. when there are available conversions through base types (#341, #351) - Fixed issues encountered in samples/games/CornerBlocks (#644) - An eC enum type minus an eC enum type will give an int rather than en enum - Validating constant expressions assigned to an enum based on max value, otherwise warning as incompatible (affects bool as well) - Not applying unit destination type to operands of an operator expression if the unit type is the base unit class (e.g. Angle rather than Radians) (#771) - Added a conversion property from MinMaxValue to AnchorValue, as it is common to use size values to specify anchors values and it would   no longer go through the 'int' conversion property

+++ compiler/bootstrap/ecere/bootstrap/TempFile.c
@@ -114,7 +114,7 @@ unsigned int noExpansion;
 char *  defaultProperty;
 unsigned int comRedefinition;
 int count;
-unsigned int isRemote;
+int isRemote;
 unsigned int internalDecl;
 void *  data;
 unsigned int computeSize;
@@ -497,7 +497,7 @@ unsigned int __ecereMethod___ecereNameSpace__ecere__sys__TempFile_GetSize(struct
 {
 struct __ecereNameSpace__ecere__sys__TempFile * __ecerePointer___ecereNameSpace__ecere__sys__TempFile = (struct __ecereNameSpace__ecere__sys__TempFile *)(this ? (((char *)this) + __ecereClass___ecereNameSpace__ecere__sys__TempFile->offset) : 0);
 
-return (unsigned int)__ecerePointer___ecereNameSpace__ecere__sys__TempFile->size;
+return __ecerePointer___ecereNameSpace__ecere__sys__TempFile->size;
 }
 
 unsigned int __ecereMethod___ecereNameSpace__ecere__sys__TempFile_Truncate(struct __ecereNameSpace__ecere__com__Instance * this, unsigned int size)
@@ -507,7 +507,7 @@ struct __ecereNameSpace__ecere__sys__TempFile * __ecerePointer___ecereNameSpace_
 __ecerePointer___ecereNameSpace__ecere__sys__TempFile->buffer = __ecereNameSpace__ecere__com__eSystem_Renew(__ecerePointer___ecereNameSpace__ecere__sys__TempFile->buffer, sizeof(unsigned char) * (size));
 __ecerePointer___ecereNameSpace__ecere__sys__TempFile->size = (unsigned int)size;
 __ecerePointer___ecereNameSpace__ecere__sys__TempFile->allocated = (unsigned int)size;
-if(__ecerePointer___ecereNameSpace__ecere__sys__TempFile->position > (unsigned int)size)
+if(__ecerePointer___ecereNameSpace__ecere__sys__TempFile->position > size)
 __ecerePointer___ecereNameSpace__ecere__sys__TempFile->position = (unsigned int)size;
 return 0x1;
 }
